requirements.txt
asgiref==3.11.0
Django==6.0
django-ckeditor==6.7.3
django-js-asset==3.1.2
pillow==12.0.0
sqlparse==0.5.5
